% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_root_loop.R
\name{sqrt_newton}
\alias{sqrt_newton}
\title{Function to compute the square root of a number}
\usage{
sqrt_newton(a, init, eps = 0.01, iter = 100)
}
\arguments{
\item{a}{the number whose square root is computed}

\item{init}{an initial guess}

\item{eps}{*optional* the precision. Default value: 0.01}

\item{iter}{*optional* the number of iteration. Default value: 100}
}
\description{
This function computes the square root of a number using a loop.
}
\examples{
sqrt_newton(81,2)
}
